requirements.txt
python-dotenv
spotipy
yt-dlp
ffmpeg-python
mutagen
colorama